SOFT: Extract EC functions to mech_openssl.c
Extract the OpenSSL based EC keygen, sign, verify, and derive functions into
file mech_openssl.c, so that they can be used by other tokens as well.
